extract all styles from register components
