+++ src/components/register/register.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react'
 import { View, Text, StyleSheet, Image, StatusBar, } from 'react-native';
 import RegisterForm from './register_form'
+import styles from '../../../src/styles/register/registerStyles'
 
 export default class Register extends Component {
   static navigationOptions = {
@@ -21,33 +22,3 @@ export default class Register extends Component {
     )
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#16a085"
-  },
-  logoContainer: {
-    alignItems: 'center',
-    flexGrow: 1,
-    justifyContent: 'center'
-  },
-  logo: {
-    width: 100,
-    height: 100
-  },
-  header: {
-    color: '#FFF',
-    marginTop: 5,
-    textAlign: 'center',
-    opacity: 0.7
-  },
-  info: {
-    color: '#FFF',
-    marginTop: 25,
-    textAlign: 'center',
-    width: 250,
-    opacity: 0.7,
-    fontSize: 20
-  }
-})
